+++ backend/app/core/__init__.py
@@ -1,1 +1,1 @@
-Core config: settings,security,database connection
+#Core config: settings,security,database connection
