+++ backend/app/models/__init__.py
@@ -1,1 +1,15 @@
 #SQlAlchemy ORM models will be defined here
+
+from .Employee import Employee
+from .attendance import Attendance
+from .Report import Report
+from .user import User
+from .Department import Department
+
+__all__ = [
+    "Employee",
+    "Attendance",
+    "Report",
+    "User",
+    "Department",
+    ]
